Display error message on login fail, update views
