+++ Studs/Login/ViewModels/LoginViewModel.swift
@@ -13,7 +13,15 @@ final class LoginViewModel {
 
     // MARK: Properties
 
+    private var errorMsg = "" {
+        didSet {
+            self.onErrorMsgChange?(self.errorMsg)
+        }
+    }
     private let disposeBag = DisposeBag()
+
+    // What should happen when the errorMsg is changed
+    var onErrorMsgChange: ((String) -> Void)?
 
     // MARK: Public methods
 
@@ -23,8 +31,15 @@ final class LoginViewModel {
         Http.login(email: email, password: password).subscribe(onNext: { [weak self] response in
             //guard let self = self else { return }
 
+            print("Response:")
             print(response)
+            
+            if response.error != nil {
+                self?.errorMsg = response.error!
+                return
+            }
         }, onError: { error in
+            print("Error:")
             print(error)
         }).disposed(by: disposeBag)
     }
